chore: add debug more statements

+++ runner/src/bench/raw.rs
@@ -6,12 +6,12 @@ pub async fn run_raw_exchange<S>(tls_stream: &mut S, payload_bytes: u32) -> miet
 where
     S: AsyncRead + AsyncWrite + Unpin,
 {
-    write_request(tls_stream, u64::from(payload_bytes))
+    write_request(tls_stream, payload_bytes)
         .await
         .into_diagnostic()
         .context("write request failed")?;
 
-    read_payload(tls_stream, u64::from(payload_bytes))
+    read_payload(tls_stream, payload_bytes)
         .await
         .into_diagnostic()
         .context("read payload failed")?;
